todo_pgsql/requirements.txt
fastapi
uvicorn
psycopg2-binary
pyjwt
python-multipart
pwdlib[argon2]
python-dotenv